solved gpu memory problem

+++ models/encoder/unidepth_encoder.py
@@ -80,7 +80,7 @@ class UniDepthExtended(nn.Module):
         outputs_gauss = {}
         # print(inputs[("K_src", 0)].shape)
         outputs_gauss[("K_src", 0)] = inputs[("K_src", 0)] if ("K_src", 0) in inputs.keys() else depth_outs["intrinsics"]
-        outputs_gauss[("inv_K_src", 0)] = torch.linalg.inv(outputs_gauss[("K_src", 0)])
+        outputs_gauss[("inv_K_src", 0)] = torch.linalg.inv(outputs_gauss[("K_src", 0)].float())
 
         if self.cfg.model.backbone.depth_cond:
             # division by 20 is to put depth in a similar range to RGB
